thirsty_truck/thirsty_truck_requirements.txt
caf.toolkit >= 0.0.7
tqdm >= 4.64.1
geopandas >= 0.12.2
matplotlib >= 3.6.2
pandas >= 1.5.2
fiona == 1.9.1
pyogrio == 0.5.1
shapely >= 2.0.1
bokeh >= 3.1.0
openpyxl >= 3.0.10